+++ src/app/resume/components/formSection.module.css
@@ -108,31 +108,6 @@
   padding-bottom: 1px;
 }
 
-.inputsContainer .textareaBox {
-  width: 100%;
-  height: 198px;
-}
-
-.textarea {
-  width: 100%;
-  max-width: 100%;
-  height: 100%;
-  max-height: 100%;
-
-  background: #ffffff50;
-  border: 1px solid #ffffff95;
-  border-radius: 10px;
-  resize: none;
-  padding: 15px;
-
-  font-family: "nanum-squere";
-  color: var(--dark-green);
-}
-
-.textarea::placeholder {
-  color: var(--dark-green);
-}
-
 .resultBtn {
   position: absolute;
   bottom: 100px;
