+++ src/app/resume/components/sectionForm.module.css
@@ -1,0 +1,33 @@
+.radioInputBox {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+
+  gap: 15px;
+}
+
+.textareaBox {
+  width: 100%;
+  height: 198px;
+}
+
+.textarea {
+  width: 100%;
+  max-width: 100%;
+  height: 100%;
+  max-height: 100%;
+
+  background: #ffffff50;
+  border: 1px solid #ffffff95;
+  border-radius: 10px;
+  resize: none;
+  padding: 15px;
+
+  font-family: "nanum-squere";
+  color: var(--dark-green);
+}
+
+.textarea::placeholder {
+  color: var(--dark-green);
+}
